requirements.txt
altgraph==0.17.4
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.4
MouseInfo==0.1.3
numpy==1.26.2
opencv-python==4.8.1.78
packaging==23.2
pefile==2023.2.7
Pillow==10.1.0
plyer==2.1.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
pyinstaller==6.2.0
pyinstaller-hooks-contrib==2023.10
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
python-dotenv==1.0.0
pytweening==1.0.7
pywin32-ctypes==0.2.2
requests==2.31.0
urllib3==2.1.0
